+++ setup.py
@@ -6,7 +6,7 @@ with open("README.md", encoding="utf-8") as fh:
     long_description = fh.read()
 
 setup(
-    name="umx",
+    name="openunmix",
     version=umx_version,
     author="Fabian-Robert Stöter",
     author_email="fabian-robert.stoter@inria.fr",
