+++ tests/test_transforms.py
@@ -1,7 +1,7 @@
 import pytest
 import numpy as np
 import torch
-import model
+from openunmix import model
 
 
 @pytest.fixture(params=[4096, 4096*10])
